fix: edit applicant data hierarchy change
